students/K3339/Kramar_Kirill/Lr1/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.34.0
sqlmodel==0.0.38
SQLAlchemy==2.0.49
alembic==1.14.1
python-dotenv==1.0.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psycopg[binary]==3.2.13
